# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_transport_rest_handler.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: Invalid input
        type: string
      message:
        example: The provided data is invalid
        type: string
    type: object
  internal_transport_rest_handler.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
    type: object
  sarc-ng_internal_domain_building.Building:
    properties:
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  sarc-ng_internal_domain_class.Class:
    properties:
      capacity:
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  sarc-ng_internal_service_building.CreateBuildingDTO:
    properties:
      code:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
  sarc-ng_internal_service_building.UpdateBuildingDTO:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - code
    - id
    - name
    type: object
  sarc-ng_internal_service_class.CreateClassDTO:
    properties:
      capacity:
        minimum: 1
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  sarc-ng_internal_service_class.UpdateClassDTO:
    properties:
      capacity:
        minimum: 1
        type: integer
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: SARC Next Generation - Resource Reservation and Management API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SARC-NG API
  version: "1.0"
paths:
  /buildings:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all buildings in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of buildings
          schema:
            items:
              $ref: '#/definitions/sarc-ng_internal_domain_building.Building'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Get all buildings
      tags:
      - buildings
    post:
      consumes:
      - application/json
      description: Create a new building with the provided name and code
      parameters:
      - description: Building creation data
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/sarc-ng_internal_service_building.CreateBuildingDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created building
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_building.Building'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Create a new building
      tags:
      - buildings
  /buildings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a building by its ID
      parameters:
      - description: Building ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Building deleted successfully
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.SuccessResponse'
        "400":
          description: Invalid building ID
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Building not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Delete a building
      tags:
      - buildings
    get:
      consumes:
      - application/json
      description: Retrieve a specific building by its unique identifier
      parameters:
      - description: Building ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Building details
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_building.Building'
        "400":
          description: Invalid building ID
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Building not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Get building by ID
      tags:
      - buildings
    put:
      consumes:
      - application/json
      description: Update an existing building's name and code by ID
      parameters:
      - description: Building ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Building update data
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/sarc-ng_internal_service_building.UpdateBuildingDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated building
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_building.Building'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Building not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Update an existing building
      tags:
      - buildings
  /classes:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all classes in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of classes
          schema:
            items:
              $ref: '#/definitions/sarc-ng_internal_domain_class.Class'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Get all classes
      tags:
      - classes
    post:
      consumes:
      - application/json
      description: Create a new class with the provided name and capacity
      parameters:
      - description: Class creation data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/sarc-ng_internal_service_class.CreateClassDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created class
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_class.Class'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Create a new class
      tags:
      - classes
  /classes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a class by its ID
      parameters:
      - description: Class ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Class deleted successfully
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.SuccessResponse'
        "400":
          description: Invalid class ID
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Class not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Delete a class
      tags:
      - classes
    get:
      consumes:
      - application/json
      description: Retrieve a specific class by its unique identifier
      parameters:
      - description: Class ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Class details
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_class.Class'
        "400":
          description: Invalid class ID
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Class not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Get class by ID
      tags:
      - classes
    put:
      consumes:
      - application/json
      description: Update an existing class's name and capacity by ID
      parameters:
      - description: Class ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Class update data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/sarc-ng_internal_service_class.UpdateClassDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated class
          schema:
            $ref: '#/definitions/sarc-ng_internal_domain_class.Class'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "404":
          description: Class not found
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_transport_rest_handler.ErrorResponse'
      summary: Update an existing class
      tags:
      - classes
...
schemes:
- http
- https
swagger: "2.0"